[Fix] #34 퀴즈 채점 결과값 계산 오류 해결

+++ app/src/main/java/com/example/guessme/ui/viewmodel/StartQuizViewModel.kt
@@ -66,7 +66,7 @@ class StartQuizViewModel @Inject constructor(
     }
 
     fun getScore(): Int {
-        return (answer!!.sum() / quizList.value!!.size) * 100
+        return ((answer!!.sum()).toDouble() / (quizList.value!!.size).toDouble() * 100).toInt()
     }
 
     fun isQuizLast(): Boolean {
